added reviewer comments
